+++ ignite/services/plugin/plugin.go
@@ -1,5 +1,5 @@
 // Package plugin implements ignite plugin management.
-// A ignite plugin is a binary which communicates with the ignite binary
+// An ignite plugin is a binary which communicates with the ignite binary
 // via RPC thanks to the github.com/hashicorp/go-plugin library.
 package plugin
 
@@ -21,13 +21,13 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/ignite/cli/ignite/config"
+	pluginsconfig "github.com/ignite/cli/ignite/config/plugins"
 	"github.com/ignite/cli/ignite/pkg/cliui"
 	cliexec "github.com/ignite/cli/ignite/pkg/cmdrunner/exec"
 	"github.com/ignite/cli/ignite/pkg/cmdrunner/step"
 	"github.com/ignite/cli/ignite/pkg/env"
 	"github.com/ignite/cli/ignite/pkg/gocmd"
 	"github.com/ignite/cli/ignite/pkg/xfilepath"
-	"github.com/ignite/cli/ignite/services/chain"
 )
 
 // pluginsPath holds the plugin cache directory.
@@ -39,7 +39,7 @@ var pluginsPath = xfilepath.Join(
 // Plugin represents a ignite plugin.
 type Plugin struct {
 	// Embed the plugin configuration
-	config.Plugin
+	pluginsconfig.Plugin
 	// Interface allows to communicate with the plugin via net/rpc.
 	Interface Interface
 	// If any error occurred during the plugin load, it's stored here
@@ -67,19 +67,17 @@ type Plugin struct {
 // If an error occurs during a plugin load, it's not returned but rather stored
 // in the Plugin.Error field. This prevents the loading of other plugins to be
 // interrupted.
-func Load(ctx context.Context, c *chain.Chain) ([]*Plugin, error) {
-	conf, err := c.Config()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
+func Load(ctx context.Context, cfg *pluginsconfig.Config) ([]*Plugin, error) {
 	pluginsDir, err := pluginsPath()
 	if err != nil {
 		return nil, errors.WithStack(err)
 	}
 	var plugins []*Plugin
-	for _, cp := range conf.Plugins {
+	for _, cp := range cfg.Plugins {
 		p := newPlugin(pluginsDir, cp)
 		p.load(ctx)
+
+		// TODO: override global plugins with locally defined ones
 		plugins = append(plugins, p)
 	}
 	return plugins, nil
@@ -98,7 +96,7 @@ func Update(plugins ...*Plugin) error {
 }
 
 // newPlugin creates a Plugin from configuration.
-func newPlugin(pluginsDir string, cp config.Plugin) *Plugin {
+func newPlugin(pluginsDir string, cp pluginsconfig.Plugin) *Plugin {
 	var (
 		p          = &Plugin{Plugin: cp}
 		pluginPath = cp.Path